+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/user/DecisionOutcomeCallbackHandler.java
@@ -9,7 +9,6 @@ import uk.gov.hmcts.reform.civil.callback.Callback;
 import uk.gov.hmcts.reform.civil.callback.CallbackHandler;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
-import uk.gov.hmcts.reform.civil.client.DashboardApiClient;
 import uk.gov.hmcts.reform.civil.model.BusinessProcess;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.service.FeatureToggleService;
@@ -18,7 +17,6 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import static uk.gov.hmcts.reform.civil.callback.CallbackParams.Params.BEARER_TOKEN;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CallbackType.SUBMITTED;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.MOVE_TO_DECISION_OUTCOME;
@@ -28,7 +26,6 @@ import static uk.gov.hmcts.reform.civil.enums.CaseState.DECISION_OUTCOME;
 @RequiredArgsConstructor
 public class DecisionOutcomeCallbackHandler extends CallbackHandler {
 
-    protected final DashboardApiClient dashboardApiClient;
     private final ObjectMapper objectMapper;
     protected final FeatureToggleService featureToggleService;
     private static final List<CaseEvent> EVENTS = Collections.singletonList(MOVE_TO_DECISION_OUTCOME);
@@ -42,11 +39,9 @@ public class DecisionOutcomeCallbackHandler extends CallbackHandler {
 
     private CallbackResponse changeState(CallbackParams callbackParams) {
         if (featureToggleService.isCaseProgressionEnabled()) {
-            String authToken = callbackParams.getParams().get(BEARER_TOKEN).toString();
             CaseData caseData = callbackParams.getCaseData().toBuilder()
                 .build();
             CaseData.CaseDataBuilder<?, ?> caseDataBuilder = caseData.toBuilder();
-            deleteNotificationAndInactiveTasks(caseData, authToken);
             caseDataBuilder.businessProcess(BusinessProcess.ready(MOVE_TO_DECISION_OUTCOME));
             return AboutToStartOrSubmitCallbackResponse.builder()
                 .state(DECISION_OUTCOME.name()).data(caseDataBuilder.build().toMap(objectMapper))
@@ -62,29 +57,4 @@ public class DecisionOutcomeCallbackHandler extends CallbackHandler {
         return EVENTS;
     }
 
-    private void deleteNotificationAndInactiveTasks(CaseData caseData, String authToken) {
-
-        dashboardApiClient.deleteNotificationsForCaseIdentifierAndRole(
-            caseData.getCcdCaseReference().toString(),
-            "DEFENDANT",
-            authToken
-        );
-
-        dashboardApiClient.makeProgressAbleTasksInactiveForCaseIdentifierAndRole(
-            caseData.getCcdCaseReference().toString(),
-            "DEFENDANT",
-            authToken
-        );
-        dashboardApiClient.deleteNotificationsForCaseIdentifierAndRole(
-            caseData.getCcdCaseReference().toString(),
-            "CLAIMANT",
-            authToken
-        );
-
-        dashboardApiClient.makeProgressAbleTasksInactiveForCaseIdentifierAndRole(
-            caseData.getCcdCaseReference().toString(),
-            "CLAIMANT",
-            authToken
-        );
-    }
 }
