+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/dashboardnotifications/claimant/ClaimantDecisionOutcomeDashboardHandlerTest.java
@@ -67,13 +67,14 @@ public class ClaimantDecisionOutcomeDashboardHandlerTest extends BaseCallbackHan
                 CallbackRequest.builder().eventId(UPDATE_DASHBOARD_TASK_LIST_CLAIMANT_DECISION_OUTCOME.name()).build()
             ).build();
 
-            HashMap<String, Object> scenarioParams = new HashMap<>();
 
-            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
             when(toggleService.isCaseProgressionEnabled()).thenReturn(true);
-
+            HashMap<String, Object> scenarioParams = new HashMap<>();
+            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
             handler.handle(params);
 
+            // Then
+            verifyDeleteNotificationsAndTaskListUpdates(caseData);
             verify(dashboardApiClient).recordScenario(
                 caseData.getCcdCaseReference().toString(),
                 SCENARIO_AAA6_CLAIMANT_TRIAL_READY_DECISION_OUTCOME.getScenario(),
@@ -92,13 +93,12 @@ public class ClaimantDecisionOutcomeDashboardHandlerTest extends BaseCallbackHan
                 CallbackRequest.builder().eventId(UPDATE_DASHBOARD_TASK_LIST_CLAIMANT_DECISION_OUTCOME.name()).build()
             ).build();
 
-            HashMap<String, Object> scenarioParams = new HashMap<>();
-
-            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
             when(toggleService.isCaseProgressionEnabled()).thenReturn(true);
-
+            HashMap<String, Object> scenarioParams = new HashMap<>();
+            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
             handler.handle(params);
-
+            // Then
+            verifyDeleteNotificationsAndTaskListUpdates(caseData);
             verify(dashboardApiClient).recordScenario(
                 caseData.getCcdCaseReference().toString(),
                 SCENARIO_AAA6_CLAIMANT_DECISION_OUTCOME.getScenario(),
@@ -118,13 +118,12 @@ public class ClaimantDecisionOutcomeDashboardHandlerTest extends BaseCallbackHan
                 CallbackRequest.builder().eventId(UPDATE_DASHBOARD_TASK_LIST_CLAIMANT_DECISION_OUTCOME.name()).build()
             ).build();
 
-            HashMap<String, Object> scenarioParams = new HashMap<>();
-
-            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
             when(toggleService.isCaseProgressionEnabled()).thenReturn(true);
-
+            HashMap<String, Object> scenarioParams = new HashMap<>();
+            when(mapper.mapCaseDataToParams(any())).thenReturn(scenarioParams);
             handler.handle(params);
-
+            // Then
+            verifyDeleteNotificationsAndTaskListUpdates(caseData);
             verify(dashboardApiClient).recordScenario(
                 caseData.getCcdCaseReference().toString(),
                 SCENARIO_AAA6_CLAIMANT_TRIAL_READY_DECISION_OUTCOME.getScenario(),
@@ -132,5 +131,18 @@ public class ClaimantDecisionOutcomeDashboardHandlerTest extends BaseCallbackHan
                 ScenarioRequestParams.builder().params(scenarioParams).build()
             );
         }
+    }
+
+    private void verifyDeleteNotificationsAndTaskListUpdates(CaseData caseData) {
+        verify(dashboardApiClient).deleteNotificationsForCaseIdentifierAndRole(
+            caseData.getCcdCaseReference().toString(),
+            "CLAIMANT",
+            "BEARER_TOKEN"
+        );
+        verify(dashboardApiClient).makeProgressAbleTasksInactiveForCaseIdentifierAndRole(
+            caseData.getCcdCaseReference().toString(),
+            "CLAIMANT",
+            "BEARER_TOKEN"
+        );
     }
 }
